requirements.txt
Crawl4AI==0.4.247
python-dotenv==1.0.1
pydantic==2.10.6
pandas==2.1.4
groq==0.4.1
supabase==2.2.0
